Hang

+++ code/tracking.h
@@ -13,7 +13,7 @@
 
 namespace substab {
 	namespace Tracking {
-		void genTrackMatrix(const std::vector<cv::Mat>& images, Eigen::MatrixXd& trackMatrix);
+		void genTrackMatrix(const std::vector<cv::Mat>& images, Eigen::MatrixXf& trackMatrix);
 	}
 }
 #endif //SUBSPACESTAB_TRACKING_H
